move app_manager.__init__ to app_manager.helpers
